Wands and Gain Ability 0.1
Wands and Gain Ability 0.1 by Grunt ( http://bilious.alt.org/?250 )

Dipping a wand of nothing into a non-cursed potion of gain ability
causes it to gain an ability, i.e. to change into a wand of something
(chosen by the normal wand probabilities). Conversely, dipping a wand
that is not a wand of nothing into a cursed potion of gain ability will
strip it of its ability, i.e. turn it into a wand of nothing. The normal
wand polymorph penalty applies - the wand will usually increment its
recharge counter. As it is not a polymorphing  this does not count
against polypileless conduct.