Fire traps removed from Plane of Fire
At this stage no player will be harmed by this type of trap and it
should reduce amount of spam messages.

http://nethackwiki.com/wiki/Plane_of_Fire

"A side effect to the numerous fire traps is that, since most of the
monsters on the Plane of Fire are immune to fire, you will practically
be swarmed with a mind-numbing myriad of messages about how "foo stepped
in a fire trap but is unharmed", potentially causing you to skip past
messages that are A Bad Idea to skip. Coupled with lots of conflict
inter-monster battle messages, this motivates some people to wear a
blindfold throughout this level."